# Bank/bank.client/src/generated/models/WeatherForecast.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bank.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface WeatherForecast
 */
export interface WeatherForecast {
    /**
     * 
     * @type {Date}
     * @memberof WeatherForecast
     */
    date?: Date;
    /**
     * 
     * @type {number}
     * @memberof WeatherForecast
     */
    temperatureC?: number;
    /**
     * 
     * @type {number}
     * @memberof WeatherForecast
     */
    readonly temperatureF?: number;
    /**
     * 
     * @type {string}
     * @memberof WeatherForecast
     */
    summary?: string | null;
}

/**
 * Check if a given object implements the WeatherForecast interface.
 */
export function instanceOfWeatherForecast(value: object): value is WeatherForecast {
    return true;
}

export function WeatherForecastFromJSON(json: any): WeatherForecast {
    return WeatherForecastFromJSONTyped(json, false);
}

export function WeatherForecastFromJSONTyped(json: any, ignoreDiscriminator: boolean): WeatherForecast {
    if (json == null) {
        return json;
    }
    return {
        
        'date': json['date'] == null ? undefined : (new Date(json['date'])),
        'temperatureC': json['temperatureC'] == null ? undefined : json['temperatureC'],
        'temperatureF': json['temperatureF'] == null ? undefined : json['temperatureF'],
        'summary': json['summary'] == null ? undefined : json['summary'],
    };
}

export function WeatherForecastToJSON(json: any): WeatherForecast {
    return WeatherForecastToJSONTyped(json, false);
}

export function WeatherForecastToJSONTyped(value?: Omit<WeatherForecast, 'temperatureF'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'date': value['date'] == null ? undefined : ((value['date']).toISOString().substring(0,10)),
        'temperatureC': value['temperatureC'],
        'summary': value['summary'],
    };
}